Move notification listener to services/network
