+++ kiosk/src/js/application_updater.js
@@ -2,7 +2,7 @@ const app = require('electron').app;
 const autoUpdater = require("electron-updater").autoUpdater;
 const CurrentScreenSettingsManager = require('./current_screen_settings_manager');
 const cron = require('node-cron');
-const NotificationListener = require('./helpers/notification_listener');
+const NotificationListener = require('./services/network/notification_listener');
 const PropertiesLoader = require('./helpers/properties_load_helper');
 const UserInteractionsManager = require('./user_interactions_manager');
 const NetworkErrorsHandlingService = require('./services/error-handlers/network_errors_handling_service');
